+++ src/controller/access.js
@@ -1,8 +1,12 @@
-const { sendMessage } = require("../controller/logsSlack")
+const mailer = require("../module/mailer")
 const moment = require("moment")
 const jwt = require("jsonwebtoken")
 const bcrypt = require("bcrypt")
 const modelUser = require("../model/user")
+const crypto = require("crypto")
+const path = require("path")
+const { sendMessage } = require("../controller/logsSlack")
+
 
 async function login (req, res) {
 
@@ -54,11 +58,92 @@ async function login (req, res) {
 
 }
 
-function forgotPassword(req, res) {
+async function sendEmail(req, res) {
 
+    const { email } = req.body
+
+    try {
+
+        const user = await modelUser.findOne({ email: email })
+        const URL = "http://localhost:4000/"
+
+    if(!user){
+
+        return res.status(400).json({ status:"User not find!" })
+    }
+
+        const token = crypto.randomBytes(20).toString("hex")
+        const userName = user.userName
+
+        const now = new Date()
+        now.setHours(now.getHours()+1)
+
+        await modelUser.findByIdAndUpdate(user._id, {
+            "$set": {
+                "passwordResetToken": token,
+                "passwordResetExpires": now
+            }
+        })
+
+        mailer.sendMail({
+            to: "andersonjulio15@gmail.com",
+            from: "anderson_julio_15@hotmail.com",
+            template: "auth/forgot_Password",
+            subject:"Alteração de senha - Kipiai",
+            context: { token, userName, URL }
+        }, (err,res) => {
+            if (err){
+                console.log(err)
+            }
+        })
+
+        return res.status(200).json({ status:"E-mail has sended!" })
+            
+
+    } catch (error) { 
+        console.log(error)
+        console.log("Error on forgot email!");
+    }
     
 }
 
+async function verifyTokenLink(req,res) {
+    const reciviedToken = req.query.token
+    const user = await modelUser.findOne({ passwordResetToken:reciviedToken })
+    const now = new Date()
+
+    console.log(now);
+
+    if (!user){
+        return res.status(400).json({status:"Invalid link"})
+    }
+
+    if (user.passwordResetExpires > now ){
+
+        return res.status(400).json({status:"Link expirado"})
+
+    }
+
+    return res.status(200).json({staus:"Link validado"})
+
+}
+
+async function changePassword(req, res) {
+    const reciviedToken = req.query.token
+    const { pass } = req.body
+
+    await model.findByIdAndUpdate(user._id, {
+        "$set": {
+            "pass": pass
+        }
+    })
+    return res.status(200).json({status:"Senha alterada"})
+}
+
 module.exports = {
-    login   
+    login,
+    sendEmail,
+    changePass,
+    changePassword,
+    verifyTokenLink
 }
