+++ src/resources/mail/auth/forgot_Password.html
@@ -101,7 +101,7 @@
                             <td>
                                 <h2> Alteração de senha</h2>
                                 <p>
-                                    Olá {{ userName }} para recuperar acesso a conta realize o acesso pelo link <a href="http://localhost:4000/change_password/">http://localhost:4000/change_password/{{token}}</a>
+                                    Olá {{ userName }} para recuperar acesso a conta realize o acesso pelo link <a href="{{ URL }}change_password/?token={{ token }}">{{ URL }}change_password/?token={{token}}</a>
                                 </p>
                             </td>
                         </tr>
